fix(dsp): stop shifting negative values in offset packing and scale expressions (#365)
Thirteen packing sites and seventy scale sites shifted signed values
into undefined behaviour on the Cortex-M4 DSP path and beyond.

Offset packing: the eight sites from the report plus four identical
siblings in the s8 mul kernels, all under ARM_MATH_DSP, packed
negative offsets with (offset << 16U); now
(int32_t)(((uint32_t)off << 16) | ((uint32_t)off & 0xFFFFu)),
bit-identical over the full int32 domain (verified exhaustively by two
independent harnesses) and preserving halfword layout into the
SADD16/SMULBB consumers -- the #343 lesson. A thirteenth site,
arm_mul_scalar_s16.c packing a VALUE rather than an offset (<< 16,
0xFFFF -- a spelling both greps missed), was caught in review; unlike
the twelve it compiles on host, silenced only by the shift-base mask.

Scale: all 70 value * (1 << left_shift) sites across #355's fourteen
files became value * (int32_t)((uint32_t)1 << left_shift) -- defined at
shift 31, wrapping identically everywhere the old form was defined in
practice (1.18M-pair sweep, ls 0..31 x edge values, plus UBSan
old-traps/new-clean).

Verified in final-gate review: 102/102 objects byte-identical on the
shipped compiler across m0/m4/m55, and the DSP path executed under
unmasked UBSan via an intrinsics shim -- main traps at the packing
site, this head does not.

Still open on #357: left_shift bounding, mask narrowing (blocked on
shared inline helpers), and the PACK_Q15x2_32x1 / PACK_S8x4_32x1 macro
pair. Refs #357.

+++ Source/BasicMathFunctions/arm_add_scalar_s16.c
@@ -99,14 +99,14 @@ arm_cmsis_nn_status arm_add_scalar_s16(const int16_t *input_1_vect,
     int16_t sum_1, sum_2;
     int32_t loop_count = block_size / 2;
 
-    input_1 = input_1_vect[0] * (1 << left_shift);
+    input_1 = input_1_vect[0] * (int32_t)((uint32_t)1 << left_shift);
     input_1 = arm_nn_requantize(input_1, input_1_mult, input_1_shift);
 
     while (loop_count > 0)
     {
         two_halfword_2 = arm_nn_read_q15x2_ia(&input_2_vect);
 
-        input_2 = (int16_t)(two_halfword_2 & 0xFFFF) * (1 << left_shift);
+        input_2 = (int16_t)(two_halfword_2 & 0xFFFF) * (int32_t)((uint32_t)1 << left_shift);
         input_2 = arm_nn_requantize(input_2, input_2_mult, input_2_shift);
         sum = input_1 + input_2;
         sum = arm_nn_requantize(sum, out_mult, out_shift);
@@ -114,7 +114,7 @@ arm_cmsis_nn_status arm_add_scalar_s16(const int16_t *input_1_vect,
         sum = MIN(sum, out_activation_max);
         sum_1 = (int16_t)sum;
 
-        input_2 = (int16_t)(two_halfword_2 >> 16) * (1 << left_shift);
+        input_2 = (int16_t)(two_halfword_2 >> 16) * (int32_t)((uint32_t)1 << left_shift);
         input_2 = arm_nn_requantize(input_2, input_2_mult, input_2_shift);
         sum = input_1 + input_2;
         sum = arm_nn_requantize(sum, out_mult, out_shift);
@@ -131,7 +131,7 @@ arm_cmsis_nn_status arm_add_scalar_s16(const int16_t *input_1_vect,
     while (loop_count > 0)
     {
         /* C = A + B */
-        input_2 = *input_2_vect++ * (1 << left_shift);
+        input_2 = *input_2_vect++ * (int32_t)((uint32_t)1 << left_shift);
         input_2 = arm_nn_requantize(input_2, input_2_mult, input_2_shift);
 
         sum = input_1 + input_2;
